+++ Package.swift
@@ -11,3 +11,14 @@ let package = Package(
     ],
     swiftLanguageVersions: [4]
 )
+
+#if os(Linux)
+    package.dependencies = [.package(url: "https://github.com/stephencelis/CSQLite.git", from: "0.0.3")]
+    package.targets = [
+        .target(name: "SQLite", exclude: ["Extensions/FTS4.swift", "Extensions/FTS5.swift"]),
+        .testTarget(name: "SQLiteTests", dependencies: ["SQLite"], path: "Tests/SQLiteTests", exclude: [
+            "FTS4Tests.swift",
+            "FTS5Tests.swift"
+        ])
+    ]
+#endif