+++ Sources/SQLite/Core/Statement.swift
@@ -26,7 +26,9 @@
 import sqlite3
 #elseif SQLITE_SWIFT_SQLCIPHER
 import SQLCipher
-#elseif SWIFT_PACKAGE
+#elseif os(Linux)
+import CSQLite
+#else
 import SQLite3
 #endif