+++ Tests/SQLiteTests/TestHelpers.swift
@@ -69,7 +69,7 @@ class SQLiteTestCase : XCTestCase {
 
     func async(expect description: String = "async", timeout: Double = 5, block: (@escaping () -> Void) -> Void) {
         let expectation = self.expectation(description: description)
-        block(expectation.fulfill)
+        block({ expectation.fulfill() })
         waitForExpectations(timeout: timeout, handler: nil)
     }